+++ elide-async/src/main/java/com/yahoo/elide/async/service/dao/DefaultAsyncAPIDAO.java
@@ -3,7 +3,7 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.async.service;
+package com.yahoo.elide.async.service.dao;
 
 import static com.yahoo.elide.core.EntityDictionary.NO_VERSION;
