+++ elide-async/src/main/java/com/yahoo/elide/async/service/thread/AsyncAPICancelThread.java
@@ -3,7 +3,7 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.async.service;
+package com.yahoo.elide.async.service.thread;
 
 import static com.yahoo.elide.core.EntityDictionary.NO_VERSION;
 
@@ -11,6 +11,7 @@ import com.yahoo.elide.Elide;
 import com.yahoo.elide.async.models.AsyncAPI;
 import com.yahoo.elide.async.models.AsyncQuery;
 import com.yahoo.elide.async.models.QueryStatus;
+import com.yahoo.elide.async.service.dao.AsyncAPIDAO;
 import com.yahoo.elide.core.DataStoreTransaction;
 import com.yahoo.elide.core.Path.PathElement;
 import com.yahoo.elide.core.RequestScope;
