+++ elide-async/src/main/java/com/yahoo/elide/async/service/thread/AsyncAPIUpdateThread.java
@@ -3,12 +3,13 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.async.service;
+package com.yahoo.elide.async.service.thread;
 
 import com.yahoo.elide.Elide;
 import com.yahoo.elide.async.models.AsyncAPI;
 import com.yahoo.elide.async.models.AsyncAPIResult;
 import com.yahoo.elide.async.models.QueryStatus;
+import com.yahoo.elide.async.service.dao.AsyncAPIDAO;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
